Add an isAtom type guard util function to ensure we have an atom (which could be of type Atom, string, or undefined.
This helps tighten the type at runtime and improves readability.
